A New Way
We are removing choice from the user.
If you want to control the port, use the `--publish` switch.
The only allowed listening interface is '127.0.0.1'.
Since we are not allowing anything other than localhost, we are
disabling authentication.

The user must specify
* a bin url - a url to a git repo containing bin items
  put your scripts in here and they will be available to you
* an upstream url - this will determine the name of your project
  (so you don't have to specify it)
* an origin url

The user probably should provide a .ssh volume

The container will create a workspace directory with git remotes
upstream and origin.